Verify error message contains path before parsing
When trying to import a module that doesn't exist, elm-make returns an error that doesn't contain a path. That caused the regex exec to return null, and destructuring to fail. At that point the plugin throws an exception, and the actual root cause is masked. This prevents that failure mode, allowing the message to display unaltered if the path regex doesn't match.